feat: Expose Compartment.evaluate() via dev console
Adds `kernel.evaluate()` method to dev console, which causes the
provided source string to be evaluated within a Compartment in an iframe
attached to the offscreen document. This is the same approach used to
evaluate Snaps in the MetaMask extension, as well as Google's
recommended way to use `eval()` as an MV3 extension.