feat(streams): Add error handling
The streams originally did not have any error handling capabilities,
since the underlying MessagePort was assumed to be completely reliable.
This was a bad asssumption, because `MessagePort.postMessage()` does in
fact throw DOMException errors under various conditions. These
conditions are not necessarily plausible, but we should handle the
possibility nevertheless.

Therefore, the streams now include error handling, and throwing a writer
will cause the reader on the opposite side to throw as well.