fix(shims): Remove build command from shims test
It was supposed that, because a critical behavior of the shims is
their behavior under bundling, the test command in the shims
package should include building prior. However, it is worth noting
that the build is assumed to run in node 20 and fails in node 18
due to the nonexistence of import.meta.dirname at that version.

However, the hard requirement is only that the build be performed
prior to testing. The CI does this already, building with node 20
and then testing under both 18 and 20. Thus the build prephase can
be safely removed from the test command without endangering the
consistency of the `main` branch; caveat that developers may need
to rebuild before testing while developing in the shims package.