build(extension): Cooperatively correct jsTrustedPrelude implementation.
The implementation throws errors only if its configuration assumptions are violated.
Otherwise, it proactively externalizes trusted preludes and attempts to correct any
correctness-breaking import permutations that may have occurred during bundling,
merely warning that some corrective action was necessary but not breaking the build.

The configuration assumptions are as follows.
 - If a file has a declared trusted prelude, the file must be declared an entry point.
 - If a file has a declared trusted prelude, it must not transitively import more than
     one declared trusted prelude.

The result is a plugin which ensures that a declared trusted prelude file is never
imported after any other code, bandaging bundles which unintentionally violate this
constraint and rejecting source code which intentionally violates this constaint.