feat(streams): Add error marshaling
Adds error marshaling to the `BaseStream` classes such that native
errors are serializes before dispatched and unserialized before yielded
on reads. Foregoes the use of `@endo/marshal` for now, since it doesn't
support e.g. `error.cause`. The marshaling functions can be
parameterized to the stream classes in the future.