feat(streams)!: Synchronize duplex streams
Introduces a notion of "synchronization" to duplex streams, such that
they must receive an acknowledgement from their remote counterpart
before they can beging processing reads and writes. This is similar to
the `cork`/`uncork` dyad in Node.js streams. The protocol recapitalutes
the one used for the same purpose in `@metamask/post-message-stream`,
which is itself modeled on the TCP SYN / SYN-ACK / ACK handshake.