refactor(ocap-kernel): Remove `waitForSyscallsToComplete()` (#527)
We added `VatSyscall.waitForSyscallsToComplete()` with a view to
ensuring that `VatHandles` could only perform certain operations at the
end of the crank. However, while syscalls are asynchronous from the
perspective of the vat worker, they are handled synchronously by the
kernel, wherefore `waitForSyscallsToComplete()` is extraneous. (Whenever
that method is called, there will be no pending syscalls.)