feat: Bundle iframe and vat worker via `kernel-browser-runtime` (#531)
#522 introduced the `kernel-browser-runtime` package, which bundled the
kernel Web Worker. Further experiments indicated that it would be
salutary for this package to bundle the vat worker, i.e. vat iframe, as
well.

To facilitate this, we replace `esbuild` with `vite` in
`kernel-browser-runtime`, and extract all vat iframe-related
functionality to this package. We also extract the extension's Vite
plugins into a dedicated, private package so that they can be reused.

Finally, during development, it was discovered that Vite duplicates the
`@sqlite.org/sqlite.wasm` WASM binaries. This was a preexisting issue in
the extension, when we bundled the kernel worker there. Thankfully, the
duplicate files are extraneous and can be deleted, which is done using
an inline plugin in `kernel-browser-runtime`.

Many red herrings died in the creation of this PR.