feat: Make export paths compatible with Browserify (#533)
Ref: #461 

The MetaMask extension is built for production using Browserify. Due to
its age, Browserify ignores the `exports` field of `package.json`.
Instead, when it encounters an import like `@metamask/streams/browser`,
it will search the root directory of that package in `node_modules` for
a file like `browser.js`. This PR adds such "mirror files" to all of our
`@metamask/*` packages with export paths. For prior art, see:
https://github.com/MetaMask/providers/pull/351

In addition, we consolidate Vite's "assets" into the root output
directory in `kernel-browser-runtime`. This turned out not be a
necessary change, but I like it nevertheless. (The single file in that
directory isn't an "asset", it's just a `.js` file.)