feat(logger): Support streaming falsey values (#542)
Provides unofficial support for the passing of falsey values through the
streamLogger. The officially supported type signature of the logger
aliases (log, info, warn, debug, error) expects the first argument to be
a string.
```ts
type LogAlias = (...args: [] | [string, ...unknown[]]) => void;
```
But in the pure javascript contexts of interest (i.e. vat bundles), it
is easier just to support non-string values than to attempt to surface
an error to the user. A gotchya here is falsey values did not previously
pass the serialization round trip, which this PR addresses.