feat: Add kernel service object support (#563)
This PR adds support for kernel-privlleged (i.e., syscall-like)
capabilities that can be messaged from vats just like any other imported
object reference. While motivated by the need to provide a way for vats
to access the recently added or soon to be added revocation and ocap URL
manipulation services, it is specifically designed to be extensible, so
that new such capabilities can be added in the future without
modification to the kernel itself.

A new kernel method, `registerKernelServiceObject(name: string, service:
object): void` allows code hosting a kernel to install such services and
register them by name. This must be called for a service prior to
loading any subclusters that would make use of it. The service itself is
created as a regular `Far` object (basically, an object whose only own
properties are functions, i.e., methods) that can have whatever methods
you like. When invoked, parameters are deserialized using the
`kernel-marshal` package, so values that are exported object references
will be tokens that you can pass as parameters in subsequent messages
that the kernel service sends (using the kernel's `queueMessage` method)
or you can interrogate their associated kref with the `krefOf` function
from the `kernel-marshal` package (the tests here give an example of the
latter). The return value from a service method resolves the promise
that the requesting vat holds from the original message send.

The cluster configuration object has been augmented with a new optional
`services` property, whose value is an array of the names of any
registered service capabilities you wish to make available to the
subcluster being configured. These services will be passed in a new
second parameter to the `bootstrap` method. This parameter will be an
object whose property names are the service names and whose property
values are the corresponding service objects. The bootstrap vat can use
these itself, or pass them on to other vats in the cluster, just as it
would with the references to vat root objects that are passed in the
first parameter to `bootstrap`. Vats can send messages to these objects
in the normal way.