fix: switch to Vite only for kernel-ui build (#567)
Previously we used `ts-bridge` together with Vite because `ts-bridge`
did not emit style files. This approach broke sourcemaps and caused the
build step of the [MetaMask extension to
fail](https://github.com/MetaMask/metamask-extension/pull/34029). With
this change we rely solely on Vite to bundle code and assets. Vite now
emits `styles.css` and preserves correct source maps. This simplifies
the build and fixes the extension build error.