fix: Remove redundant kernel promise ref count increment (#565)
While doing preliminary research to begin expanding our integration test
suite in `packages/kernel-test`, we discovered that the refcounts for an
object in `garbage-collection.test.ts` seemed too high by a count of 1
or 2. As it turned out, they were definitely at least off by one,
because we were redundantly incrementing the refcounts of kernel promise
data slots in `KernelStore.resolveKernelPromise()`. These are already
incremented in `KernelQueue.resolvePromise()`, which calls
`resolveKernelPromise()`. With this change, we keep the former and
abolish the latter.

We also include some changes that were made in the course of debugging,
that may prove generally helpful. This includes:
- A handful of `console.debug()` calls that help when debugging
refcounts.
- Logging an error if `Kernel.reset()` fails.

Finally, we remove some incidental tooling cruft from `kernel-test`.