feat(ocap-kernel): Throw if subcluster launch fails (#566)
Closes: #571 

Previously, the invocation which launched a subcluster would assume
success in many cases where it ought not. This PR causes the promise
returned by `launchSubcluster` to reject if an error occurs in any of
the following places.
- In the global namespace of a vat in the cluster
- During the buildRootObject call of any vat in the cluster
- During the bootstrap call of the cluster's bootstrap vat, if one was
designated.

It is also possible that an uncaught promise rejects in any of the above
scopes. Although the promise specification may not consider this to be
an error condition, vat developers may nonetheless appreciate this
scenario surfacing to them in some way. This PR adds placeholder tests
for detecting these situations, but does not implement them or the
functionality they imply.