fix: Restore `kernel-browser-runtime` sourcemaps in extension (#575)
When we introduced the `kernel-browser-runtime` package, we effectively
broke debugging the kernel by breaking its sourcemaps. This happened
because we copied its build output (also from Vite) using
`vite-plugin-static-copy`. To fix this, this PR modifies the extension
`vite.config.ts` to build `kernel-browser-runtime` directly.

A bigger mystery is why Vite doesn't pull in the sourcemaps of bundled
dependencies that ship them. The symptom of this is that we see
transpiled `.js` and `.mjs` files in the debugger rather than original
sources. However, this is still a significant improvement over the
status quo, and we can address the larger sourcemap problem later.