feat: Add kernel language model service (#586)
Lays out a strategy for language model integration into the ocap-kernel.

## Changes
- **New package**: `@ocap/kernel-language-model-service`.
Implementation-specific dependencies are isolated into distinct exports.
### `@ocap/kernel-language-model-service`
- Provides a streaming language model interface. There is a
`makeInstance` method, which returns an object with a `sample` method.
The name sample is chosen to align with the semantics of `sample` in the
model context protocol.
- Provides a nodejs ollama client implementation, `OllamaNodejsService`.
- Adds an e2e test which can be used to connect to and test integration
with an ollama server running locally on the development workstation.