feat(cli): Make `ocap bundle` throw on failure (#606)
Previously, `bundle` command would only log its errors on failure,
without failing. This PR ensures that it throws the error and causes the
CLI to exit with a non-zero error code. Also ensures that a new
temporary directory is created for each run of the bundle tests, and
that it is deleted at the end of the tests.