First pass of support for kernel-kernel network comms (#578)
This PR represents the first installment of support for kernel-to-kernel
network communications, enabling user code in a vat running in one
kernel to exchange message traffic with user code running in a vat in a
different kernel on a different machine elsewhere on the Internet.

I've labelled this "the first installment" because a number of things
are still noticeably lacking, but we'd like to get this code into the
main branch so people can start building on it.

Missing things (a non-exclusive list):
- Any kind of automated tests. It should be possible to write tests for
a lot of the vat-talking-to-a-remote-vat stuff by mocking both the
remote vat and the networking layer, but exactly how to write
integration tests for the network layer itself remains elusive.
- Clean support for the persistent state associated with remote
connections. Mostly this is because right now the comms code doesn't
make a clean distinction between restarting a kernel from persistent
state and starting a kernel afresh, which is in turn mostly because
there's some hackery to make debugging and (manual) testing a lot easier
wherein some state that should be persistent across restarts isn't and
some other state that shouldn't be persistent across restarts is.
- Some of the handling of ocap URLs is only rudimentary; in particular,
the connection setup code doesn't make good use of routing hints in a
URL.
- As I've previously mentioned in standups, the uses of cryptography
here are amateurish and need a pass by someone who is qualified to do
that. I've posted a writeup of that to Slack for Erik to pass along to
the relevant other bodies in the company.
- The "network backbone", such as it is, is a complete placeholder,
making use of a single relay server that's running on a server next to
my desk hosted at `fudco.com`, with network addresses referring to that
embedded in the existing source code. That's gotta go, but, as with many
of the other placeholders, it's super convenient for testing.
- Pretty much only the happy path has been exercised. There's almost
certainly a lot of weakness in error handling and recovery.