refactor: Convert package template to an actual package (#631)
Closes #628 

To reduce our manual maintenance burden and make creating new packages
easier, converts the package template used by `create-package` to an
actual template package. Specifically, the original template package,
which was not in fact a package, was turned into an actual package by
means of `create-package`, which is now used by `create-package`, which
no longer maintains a template package in its source files.