Make `Logger` obey log level settings (#703)
With this change, log messages can be suppressed if they are at a lower
level than whatever the current log level setting is (i.e., if the
setting is `'error'`, a call to `log.debug(...)` will output nothing).

Currently the log level is set on on the console transport via an argument to `makeConsoleTransport()`.

For backwards compatibility until we decide otherwise, if unspecified
the level defaults to `'debug'`, i.e., always print all log messages.
However, now that we can obey the level setting, we should probably
consider changing the default. We probably should also consider adding
an `'off'` level that suppresses all log output, but for weird
complicated reasons having to do with conflation of the `console` API
with the level name strings that's tricky. (I'd argue that the default
level should either be `'error'` or `'off'`.)

Eventually we will probably also want to read the setting from an
environment variable, but that's an additional bundle of complication
that deserves its own issue.

Closes #679