Refactor remote location hint handling (#712)
This set of changes refactors how remote location hints are handled. Previously, hints were carried by each remote message send. This was concerning because the hint information is (a) potentially quite bulky, (b) highly redundant from one message to another, and (c) changes relatively rarely. The bulk is also a concern once we start persisting the remote message queues, which this is a precursor to doing (plus, once the message queues are persisted, storing hints on a per-message basis risks inconsistencies if the hint information changes while the messages are on disk). This PR adds a new request to the platform services API, `registerLocationHints`, which is used prior to sending the first message to a remote plus can be called any time the hint information is updated, and then removes the hint information from the message transmission flow entirely.