test(kernel-language-model-service): Add queue language model service for testing (#710)
This PR adds a queue language model service to the monorepo. The
implementation is intended for mocking language model behavior where a
remote API or local model execution are both unavailable. The intended
streaming response is pushed to the model prior to sampling.

Also adds tests to kernel-test which exercise the basic shape of a
LanguageModelService.